Extract Elasticsearch::Index#amend
This is generally a good idea, but in this case it enables us to
extract the tests as well.
